hwclock: check the hwclock until --set operation finishes
On some system, a HP ProLiant m400 Server in my case, this hwclock --set
operation will take about 0.5 - 1 second to finish. So if we fork another
subprocess immediately to check the output from hwclock, it will get the
timestamp before this modification.

Use utils.system_output() to wait for the hwclock --set command to finish.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>